Create README.md
This GitHub repository has been created for the research project titled "Improving aerial targeting precision: a study on point cloud semantic segmentation with advanced deep learning algorithms."
This repository contains the code for the color-enhanced PointNet++ and RandLA-NET algorithms. Additionally, the dataset produced within the scope of the study is also provided in this repository.